refactor(engine): extract models::progress
First carve of the models.rs split (#950): the terminal progress bar is the
leaf module, with no dependency on any other future sibling. Pure move —
`pub(super)` on the four items mod.rs still uses is the only edit to a
carried line.

`include_str!` is file-relative, so the model-plan.json path in
`manifest_tests` gains one `../` for the extra directory level. It moves
here rather than with manifest.rs because mod.rs must compile its own
tests now.

Both path filters follow the file: ci.yml's `code` lane and cache-seed.yml's
push trigger now name `rust/src/models/**`, which is what the guard added in
d81672b exists to force.

Refs #950
